Move struct `transformer` into header `kernel/transformer.hpp`
